+++ InitGui.py
@@ -2,9 +2,9 @@
 # Copyright 2024 Golodnikov Sergey
 
 
+import addFC
 import FreeCAD
 import os
-import addFC
 
 
 add_base_path: str = os.path.dirname(addFC.__file__)
@@ -28,8 +28,10 @@ class addFC(FreeCAD.Gui.Workbench):
             'AddFCProperties',
             'AddFCPipe',
             'AddFCExplode',
+            'addFCInsert',
             'addFCAssistant',
         ]
+
         self.appendToolbar('addFC', self.list)
         self.appendMenu('addFC', self.list)
 
@@ -42,7 +44,17 @@ class addFC(FreeCAD.Gui.Workbench):
 
         FreeCAD.Gui.addPreferencePage(P.addFCPreferenceSpecification, 'addFC')
         FreeCAD.Gui.addPreferencePage(P.addFCPreferenceSM, 'addFC')
+        FreeCAD.Gui.addPreferencePage(P.addFCPreferenceOther, 'addFC')
+
         FreeCAD.Gui.addIconPath(add_icon_path)
+
+        # interface font:
+        conf = P.load_configuration()
+        if 'interface_font' in conf:
+            font = conf['interface_font']
+            if font[0] and font[1] != '':
+                from PySide import QtGui
+                QtGui.QApplication.setFont(QtGui.QFont(font[1], font[2]))
 
     def Activated(self): return